+++ coordinator/persistence-cassandra-5.0/pom.xml
@@ -232,7 +232,7 @@
             <Bundle-Name>Cassandra-Persistence</Bundle-Name>
             <Bundle-Description>Cassandra DB component for project</Bundle-Description>
             <Bundle-SymbolicName>io.stargate.db.cassandra_5_0</Bundle-SymbolicName>
-            <Bundle-Activator>io.stargate.db.cassandra.Cassandra50PersistenceActivator</Bundle-Activator>
+            <Bundle-Activator>io.stargate.db.cassandra.Cassandra41PersistenceActivator</Bundle-Activator>
             <Import-Package><![CDATA[
               org.slf4j,
               org.slf4j.helpers,
